requirements.txt
# Core framework dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.30.0

# LangChain for agent functionality
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Google Gemini AI
google-generativeai>=0.3.0

# Google Calendar API
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pytz